+++ src/app/features/pokemon/components/pokemon-summary/pokemon-summary.component.html
@@ -18,14 +18,6 @@
       }}
     </mat-card-subtitle>
   </mat-card-header>
-  <mat-divider></mat-divider>
-  <mat-card-content class="pokemon-types">
-    <mat-chip-set aria-label="Types">
-      @for (type of pokemonStore.selectedEntity()!.types; track type.slot) {
-        <type-chip [typeName]="type.type.name"></type-chip>
-      }
-    </mat-chip-set>
-  </mat-card-content>
   <mat-card-content class="pokemon-image">
     <img
       [ngSrc]="
@@ -40,15 +32,6 @@
       width="350"
     />
   </mat-card-content>
-  <mat-card-content class="pokemon-abilities">
-    <span class="label-large">Abilities</span>
-    <mat-chip-set aria-label="Abilities">
-      @for (ability of abilities(); track ability.slot) {
-        <mat-chip [disabled]="ability.isHidden">{{ ability.ability.names || [] | localise : 'en' : 'name' }}</mat-chip>
-      }
-    </mat-chip-set>
-  </mat-card-content>
-  <mat-divider></mat-divider>
   <mat-card-content class="pokemon-forms">
     @if (pokemonStore.entities().length > 1) {
       <h4 class="label-large">Select Form</h4>
