+++ src/app/features/pokemon/components/tabs/pokemon-details-tab/pokemon-details-tab.component.html
@@ -1,5 +1,23 @@
 <div class="details-tab">
   <div class="detail">
+    <span class="label-large">Types</span>
+    <mat-chip-set aria-label="Types">
+      @for (type of pokemonStore.selectedEntity()!.types; track type.slot) {
+        <type-chip [typeName]="type.type.name"></type-chip>
+      }
+    </mat-chip-set>
+  </div>
+  <mat-divider></mat-divider>
+  <div class="detail">
+    <span class="label-large">Abilities</span>
+    <mat-chip-set aria-label="Abilities">
+      @for (ability of abilities(); track ability.slot) {
+        <mat-chip [disabled]="ability.isHidden">{{ ability.ability.names || [] | localise : 'en' : 'name' }}</mat-chip>
+      }
+    </mat-chip-set>
+  </div>
+  <mat-divider></mat-divider>
+  <div class="detail">
     <span class="label-large">Description:</span>
     <p>{{ pokemonStore.speciesDetails().flavor_text_entries[0].flavor_text }}</p>
   </div>
