+++ src/app/features/pokemon/components/tabs/pokemon-details-tab/pokemon-details-tab.component.ts
@@ -1,10 +1,21 @@
-import { Component, inject } from "@angular/core";
+import { Component, effect, inject, signal } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { PokemonStore } from "../../../../../shared/store/pokemon.store";
 import { DecimetersToInchesPipe } from "../../../../../shared/pipes/decimetersToInches.pipe";
 import { HectogramsToPoundsPipe } from "../../../../../shared/pipes/hectogramsToPounds.pipe";
 import { MatDividerModule } from "@angular/material/divider";
 import { CaptureRatePipe } from "../../../../../shared/pipes/captureRate.pipe";
+import { LocalisePipe } from "../../../../../shared/pipes/localise.pipe";
+import { MatChipsModule } from "@angular/material/chips";
+import { Ability } from "pokenode-ts";
+import { PokemonService } from "../../../../../shared/services/pokemon.service";
+import { TypeChipComponent } from "../../../../../shared/components/type-chip/type-chip.component";
+
+interface AbilityInformation {
+  ability: Ability;
+  isHidden: boolean;
+  slot: number;
+}
 
 @Component({
   selector: "pokemon-details-tab",
@@ -14,10 +25,44 @@ import { CaptureRatePipe } from "../../../../../shared/pipes/captureRate.pipe";
     HectogramsToPoundsPipe,
     MatDividerModule,
     CaptureRatePipe,
+    LocalisePipe,
+    MatChipsModule,
+    TypeChipComponent,
   ],
   templateUrl: "./pokemon-details-tab.component.html",
   styleUrl: "./pokemon-details-tab.component.scss",
 })
 export class PokemonDetailsTabComponent {
   readonly pokemonStore = inject(PokemonStore);
+  readonly abilities = signal<AbilityInformation[]>([]);
+  readonly pokemonService = inject(PokemonService);
+
+  constructor() {
+    effect(() => {
+      if (
+        this.pokemonStore.selectedEntity()?.abilities &&
+        this.pokemonStore.selectedEntity()!.abilities!.length > 0
+      ) {
+        this.abilities.set([]);
+        this.pokemonStore.selectedEntity()!.abilities.map((pokemonAbility) => {
+          this.pokemonService
+            .getAbilityByName(pokemonAbility.ability.name)
+            .subscribe((ability: Ability) => {
+              this.abilities.update((currentAbilities) =>
+                [
+                  ...currentAbilities,
+                  {
+                    ability,
+                    slot: pokemonAbility.slot,
+                    isHidden: pokemonAbility.is_hidden,
+                  },
+                ].sort((a, b) => a.slot - b.slot)
+              );
+            });
+        });
+      } else {
+        this.abilities.set([]);
+      }
+    });
+  }
 }
