+++ src/app/features/pokemon/pages/pokemon/pokemon.component.scss
@@ -6,8 +6,10 @@
 
   .pokemon-container {
     display: flex;
-    flex-direction: column;
-    align-items: center;
+    @media (max-width: 800px) {
+      flex-direction: column;
+      align-items: center;
+    }
     gap: 20px;
     width: 100%;
     max-width: 800px;
